requirements.txt
Flask
requests
biopython
